Fundamentally changed the way the game performs saves and loads
- Saves are stored 'piece-wise' (Player, inventory, and World areas are
stored separately)
- Areas are saved temporarily as the game is played. These files are
deleted at the end of a play session or on load. On save, these
temporary files are moved over to the main save directory.
- One area at a time is now loaded and stored in the EntityManager.
Before, the game had the entire WorldMap loaded in at once, which was
bad for memory and would decrease performance the more content was
added.
- Areas are no longer defined as static entities, which was bad design
as it poorly affected the memory
- Updated .gitignore to ignore Save files (*.ser) as well as the
.project and .classpath files
- Added a few new screens to the Desert area

 Committer: Jackson Oliva <jeoliva@wpi.edu>